C_PUCTをdoubleからFloatType(float)に変更

+++ src/searcher.hpp
@@ -9,11 +9,11 @@ public:
     static bool stop_signal;
 
 protected:
-    explicit Searcher(int64_t hash_size, double C_PUCT) : hash_table_(hash_size),
-                                                          C_PUCT_(C_PUCT),
-                                                          root_index_(UctHashTable::NOT_EXPANDED),
-                                                          time_limit_(LLONG_MAX),
-                                                          node_limit_(LLONG_MAX) {}
+    explicit Searcher(int64_t hash_size, FloatType C_PUCT) : hash_table_(hash_size),
+                                                             C_PUCT_(C_PUCT),
+                                                             root_index_(UctHashTable::NOT_EXPANDED),
+                                                             time_limit_(LLONG_MAX),
+                                                             node_limit_(LLONG_MAX) {}
 
     //時間制限含め探索を続けるかどうかを判定する関数
     bool shouldStop();
@@ -33,7 +33,7 @@ protected:
     UctHashTable hash_table_;
 
     //C_PUCT
-    const double C_PUCT_;
+    const FloatType C_PUCT_;
 
     //時間
     std::chrono::steady_clock::time_point start_;
